refactor: move mockup to internal folder
