+++ packages/pigment-css-react/src/index.ts
@@ -8,4 +8,3 @@ export { default as createUseThemeProps } from './createUseThemeProps';
 export { default as internal_createExtendSxProp } from './createExtendSxProp';
 export { default as useTheme } from './useTheme';
 export { default as globalCss } from './globalCss';
-export { default as internal_styledEngineMockup } from './styledEngineMockup';
