+++ packages/pigment-css-react/src/internal/index.ts
@@ -1,0 +1,1 @@
+export { default as styledEngineMockup } from './styledEngineMockup';
